moved cross search query to a new folder
